实现方案1：用户金币字段改为DECIMAL支持小数
- 数据库迁移 V9：user.coin INT → DECIMAL(10, 2)，最小单位0.01（1分）
- User.java: Integer coin → BigDecimal coin
- UserVO.java: Integer coin → BigDecimal coin
- UserService: addCoin(int) → addCoin(BigDecimal)
- UserServiceImpl:
  * addAndCoin 使用 BigDecimal.add() 代替 + 运算符
  * addCoin 直接使用 BigDecimal 运算，避免精度丢失
- WishlistServiceImpl:
  * 移除 deposit 方法中的 intValue() 调用
  * 使用 BigDecimal.compareTo() 进行比较
  * 使用 BigDecimal.negate() 获取负值
  * abandon 方法直接使用 BigDecimal 退还金币

修复问题：
- 心愿单存入金币时精度丢失（10.5 → 10）
- 用户金币扣减与心愿单进度增加不一致
- 支持小数金币（如 0.5 金币 = 50 分）
- 类型一致性：user.coin 与 wishlist.*Amount 都是 DECIMAL

+++ src/main/java/store/scserver/my_life_app/entity/User.java
@@ -2,6 +2,7 @@ package store.scserver.my_life_app.entity;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -11,7 +12,7 @@ public class User {
     private String password;
     private Integer level;
     private Integer exp;
-    private Integer coin;
+    private BigDecimal coin;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "GMT+8")
     private LocalDateTime createdAt;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "GMT+8")
