+++ src/main/java/store/scserver/my_life_app/vo/UserVO.java
@@ -2,6 +2,7 @@ package store.scserver.my_life_app.vo;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -10,7 +11,7 @@ public class UserVO {
     private String username;
     private Integer level;
     private Integer exp;
-    private Integer coin;
+    private BigDecimal coin;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "GMT+8")
     private LocalDateTime createdAt;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "GMT+8")
